Coding-Agent (2024.8.1): Unit Test Failure Fixes (#3)
### Addressing the following unit test failures:
```json
{
    "err_msg": "Unit Test Results: 1 of 1 Failed \n\n\n### test_example_grid_with_mixed_frequency_antennas at line 17\ndef test_example_grid_with_mixed_frequency_antennas():\n        input_grid = (\n            \"............\\n\"\n            \"........0...\\n\"\n            \".....0......\\n\"\n            \".......0....\\n\"\n            \"....0.......\\n\"\n            \"......A.....\\n\"\n            \"............\\n\"\n            \"............\\n\"\n            \"........A...\\n\"\n            \".........A..\\n\"\n            \"............\\n\"\n            \"............\"\n        )\n        expected_antinodes = 14\n    \n        result = count_antinodes(input_grid)\n    \n>       assert result == expected_antinodes, (\n            f\"Failed to count correct number of antinodes.\\n\"\n            f\"Input grid:\\n{input_grid}\\n\"\n            f\"Expected {expected_antinodes} antinodes, but got {result}\"\n        )\nE       AssertionError: Failed to count correct number of antinodes.\nE         Input grid:\nE         ............\nE         ........0...\nE         .....0......\nE         .......0....\nE         ....0.......\nE         ......A.....\nE         ............\nE         ............\nE         ........A...\nE         .........A..\nE         ............\nE         ............\nE         Expected 14 antinodes, but got 0\nE       assert 0 == 14\n\nadvent_of_code/year2024/day8/part1/tests.py:37: AssertionError\n\n\n"
}
```

### Theorized solution:
```json
{
    "problem_explanation": "The `count_antinodes` function is not correctly identifying and counting the antinodes within the example grid.  The function returns 0 antinodes instead of the expected 14.",
    "optional_theorized_unit_test_fix": null,
    "optional_theorized_implementation_fix": "The issue lies within the `find_antinode` function. Currently, it calculates the coordinates of the potential antinodes by converting the x and y coordinates to integers:\n```python\nantinode_x = int(p1[0] + dx * ratio)\nantinode_y = int(p1[1] + dy * ratio)\n```\nThis conversion to integers is causing the problem, and it is not needed according to the prompt, and may introduce an off-by-one problem because it effectively rounds down to the nearest integer, so it might not be able to find antinodes precisely, potentially missing or miscalculating them. To fix this, remove the `int` conversion, and make the following change:\n```python\nantinode_x = p1[0] + dx * ratio\nantinode_y = p1[1] + dy * ratio\n```\nThis ensures that antinode coordinates are calculated as floats whenever necessary, thus allowing precise location of antinodes and also ensuring that the collinearity and distance checks accurately identify true antinodes."
}
```
### Implementation refactoring plan:
{
    "plan": [
        {
            "step": "Remove the int conversion in the find_antinode function to allow for float coordinates, ensuring precise antinode location and accurate collinearity and distance checks."
        }
    ]
}

+++ advent_of_code/year2024/day8/part1/solution.py
@@ -31,18 +31,18 @@ def are_collinear(p1: Tuple[int, int], p2: Tuple[int, int],
     area = p1[0] * (p2[1] - p3[1]) + p2[0] * (p3[1] - p1[1]) + p3[0] * (p1[1] - p2[1])
     return abs(area) < 1e-10
 
-def find_antinode(p1: Tuple[int, int], p2: Tuple[int, int]) -> List[Tuple[int, int]]:
+def find_antinode(p1: Tuple[int, int], p2: Tuple[int, int]) -> List[Tuple[float, float]]:
     """Find antinodes for two antennas of same frequency."""
     dx = p2[0] - p1[0]
     dy = p2[1] - p1[1]
-    
+
     if dx == 0 and dy == 0:
         return []
 
     antinodes = []
     for ratio in [1/3, 3]:
-        antinode_x = int(p1[0] + dx * ratio)
-        antinode_y = int(p1[1] + dy * ratio)
+        antinode_x = p1[0] + dx * ratio
+        antinode_y = p1[1] + dy * ratio
         antinodes.append((antinode_x, antinode_y))
 
     return antinodes
@@ -56,7 +56,7 @@ def count_antinodes(grid: str) -> int:
     """Count unique antinode locations within grid bounds."""
     width, height = get_grid_bounds(grid)
     antennas = get_antennas(grid)
-    antinodes: Set[Tuple[int, int]] = set()
+    antinodes: Set[Tuple[float, float]] = set()
     
     # Find antinodes for each pair of same-frequency antennas
     for freq, positions in antennas.items():
@@ -70,7 +70,7 @@ def count_antinodes(grid: str) -> int:
                         if are_collinear(p1, p2, antinode) and \
                            is_twice_distance(antinode, p1, p2):
                             antinodes.add(antinode)
-    
+                            
     return len(antinodes)
 
 def solution() -> int:
